feat: move plugin interface to sdk
