+++ go.mod
@@ -5,19 +5,22 @@ go 1.23.0
 toolchain go1.24.3
 
 require (
-	github.com/alicebob/miniredis/v2 v2.35.0
-	github.com/coder/websocket v1.8.13
-	github.com/getkin/kin-openapi v0.132.0
-	github.com/go-chi/chi/v5 v5.2.2
-	github.com/go-chi/cors v1.2.2
-	github.com/google/uuid v1.6.0
-	github.com/mark3labs/mcp-go v0.37.0
-	github.com/oapi-codegen/runtime v1.1.2
-	github.com/prometheus/client_golang v1.23.0
-	github.com/redis/go-redis/v9 v9.12.1
-	github.com/rs/zerolog v1.34.0
-	gopkg.in/yaml.v3 v3.0.1
+        github.com/alicebob/miniredis/v2 v2.35.0
+        github.com/coder/websocket v1.8.13
+        github.com/getkin/kin-openapi v0.132.0
+        github.com/go-chi/chi/v5 v5.2.2
+        github.com/go-chi/cors v1.2.2
+github.com/gaspardpetit/nfrx-sdk v0.0.0
+        github.com/google/uuid v1.6.0
+        github.com/mark3labs/mcp-go v0.37.0
+        github.com/oapi-codegen/runtime v1.1.2
+        github.com/prometheus/client_golang v1.23.0
+        github.com/redis/go-redis/v9 v9.12.1
+        github.com/rs/zerolog v1.34.0
+        gopkg.in/yaml.v3 v3.0.1
 )
+
+replace github.com/gaspardpetit/nfrx-sdk => ./nfrx-sdk
 
 require (
 	github.com/apapsch/go-jsonmerge/v2 v2.0.0 // indirect
@@ -26,7 +29,7 @@ require (
 	github.com/buger/jsonparser v1.1.1 // indirect
 	github.com/cespare/xxhash/v2 v2.3.0 // indirect
 	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
-	github.com/go-openapi/jsonpointer v0.21.0 // indirect
+github.com/go-openapi/jsonpointer v0.21.0 // indirect
 	github.com/go-openapi/swag v0.23.0 // indirect
 	github.com/invopop/jsonschema v0.13.0 // indirect
 	github.com/josharian/intern v1.0.0 // indirect
