+++ modules/llm/ext/openai/chat_completions.go
@@ -11,8 +11,8 @@ import (
 	"github.com/google/uuid"
 
 	ctrl "github.com/gaspardpetit/nfrx-sdk/contracts/control"
+	"github.com/gaspardpetit/nfrx-sdk/spi"
 	"github.com/gaspardpetit/nfrx/modules/common/logx"
-	"github.com/gaspardpetit/nfrx/modules/common/spi"
 )
 
 // ChatCompletionsHandler handles POST /api/llm/v1/chat/completions as a pass-through.
