+++ modules/llm/ext/openai/embeddings.go
@@ -16,8 +16,8 @@ import (
 	"github.com/google/uuid"
 
 	ctrl "github.com/gaspardpetit/nfrx-sdk/contracts/control"
+	"github.com/gaspardpetit/nfrx-sdk/spi"
 	"github.com/gaspardpetit/nfrx/modules/common/logx"
-	"github.com/gaspardpetit/nfrx/modules/common/spi"
 )
 
 // EmbeddingsHandler handles POST /api/llm/v1/embeddings as a pass-through.
