+++ modules/mcp/ext/mcpbroker/broker.go
@@ -13,8 +13,8 @@ import (
 
 	"github.com/coder/websocket"
 	"github.com/gaspardpetit/nfrx-sdk/contracts/mcp"
+	"github.com/gaspardpetit/nfrx-sdk/spi"
 	"github.com/gaspardpetit/nfrx/modules/common/logx"
-	"github.com/gaspardpetit/nfrx/modules/common/spi"
 	"github.com/go-chi/chi/v5"
 	"github.com/google/uuid"
 )
