+++ server/internal/ctrlsrv/metrics.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/gaspardpetit/nfrx/modules/common/spi"
+	"github.com/gaspardpetit/nfrx-sdk/spi"
 	"github.com/gaspardpetit/nfrx/server/internal/serverstate"
 )
