+++ server/templates/workerplugin/workerplugin.go
@@ -1,7 +1,7 @@
 package workerplugin
 
 import (
-	"github.com/gaspardpetit/nfrx/modules/common/spi"
+	"github.com/gaspardpetit/nfrx-sdk/spi"
 )
 
 // Plugin is a minimal example implementing plugin.Plugin and plugin.WorkerProvider.
